Compatibility with SS3.2
SS3.2 changed the way how it prepares ID values and `-` is replaced with `_`.
But to keep it backwards compatible it is better to search by `name` attribute.
Also, it is better to search for a field inside Map object, not globally. Otherwise it may return wrong field if you have more then one field in page.